requirements.txt
camelot-py
opencv-python
matplotlib
label-studio
pdf2image
pdfminer